+++ CHANGELOG.rst
@@ -7,6 +7,21 @@ Note that ``RB_ID=#`` and ``PHAB_ID=#`` correspond to associated message in comm
 Unreleased
 ----------
 
+Added
+~~~~~
+
+* finatra-http: Introduce the new streaming request and response types:
+  `c.t.finatra.http.streaming.StreamingRequest` ``PHAB_ID=D342728``,
+  `c.t.finatra.http.streaming.StreamingResponse` ``PHAB_ID=D342703``.
+  Examples are located in finatra/examples/streaming-example/.
+
+Changed
+~~~~~~~
+
+* finatra-http: Deprecated `c.t.finatra.http.response.StreamingResponse`, Use
+  `c.t.finatra.http.response.ResponseBuilder.streaming` to construct a
+  `c.t.finatra.http.streaming.StreamingResponse` instead. ``PHAB_ID=D342703``
+
 19.7.0
 ------